Refactor JSON unmarshalling code to util package
