+++ packages/nitro-util/src/index.ts
@@ -2,3 +2,4 @@ export * from './eth-client';
 export * from './math';
 export * from './go';
 export * from './hex-utils';
+export * from './json';
